[FEATURE] Show internal redirects on target page
Within big installations it is unclear if there is an internal redirect on pages. An editor can delete or hide the target page, without knowing if this page is a target for a redirect.

This patch adds an overlay icon on the page icon, to identify it as an internal redirect target, and adds a suffix to the page title in the page tree to show which redirect is using this page as a target.