[FEATURE] Add possibility on partial paths
TYPO3 is not using permalinks out of the box. When renaming the title of a page, the url will probably change as well, including from child pages. When such pages are indexed by search engines like Google, clicking such a link will result in a 404 Page Not Found.

To prevent this, this patch makes it possible to enter an old partial path which is redirected to a new path, including the child pages (if their title/URI has not changed)