[BUGFIX] Wrong type error after change to type declarations for methods
An error is thrown when an existing redirect record is edited. The error comes from the redirect repository, method getEqualRecords. The method expects as first parameter a string, but an integer is passed. The reason it expects a string is a new record uid will be a string, starting with “NEW”.

This patch changes the type for the uid parameter passed to the method